requirements.txt
py_clob_client
requests
pandas
python-dotenv
